requirements.txt
streamlit
python-dotenv
langchain
langchain-huggingface
langchain-community
google-generativeai
